requirements.txt
Flask
supabase
python-dotenv
gunicorn
Flask-Limiter
pytest
python-jose
passlib
bcrypt
Flask-CORS